% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAutoPRISMA.R
\name{getAutoPRISMA}
\alias{getAutoPRISMA}
\title{Create Prisma}
\usage{
getAutoPRISMA(prismaExampleInput, fontsize = 10, labelBuffer = 0)
}
\arguments{
\item{prismaExampleInput}{a dataframe with 1 row per item in the prima diagram.}

\item{fontsize}{how large to make the characters in the diagram}

\item{labelBuffer}{number of characters to create "buffer" around text}
}
\value{

}
\description{
Create Prisma
}
\examples{
getAutoPRISMA(prismaExampleInput)
}
