% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrisma.R
\name{getPrisma}
\alias{getPrisma}
\title{Create Prisma
These functions create a PRISMA diagram from a simple dataframe keeping track of sources and filters
It creates a DiagrammeR file run by grV}
\usage{
getPrisma(studyStatus = NULL, prismaFormat = NULL)
}
\arguments{
\item{studyStatus}{a dataframe that keeps track of}

\item{prismaFormat}{how large to make the characters in the diagram}
}
\value{
prismaTemplate A text file use to make a PRISMA diagram.
}
\description{
Create Prisma
These functions create a PRISMA diagram from a simple dataframe keeping track of sources and filters
It creates a DiagrammeR file run by grV
}
\examples{
prismaTemplate <- getPrisma(studyStatus)
prismaFormat <- getPrismaFormat(studyStatus)
prismaTemplate <- getPrisma(studyStatus, prismaFormat)
DiagrammeR::grViz(prismaTemplate)
}
