% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrismaFormat.R
\name{getPrismaFormat}
\alias{getPrismaFormat}
\title{Create Prisma
These functions create a PRISMA diagram from a simple dataframe keeping track of sources and filters
It creates a DiagrammeR file run by grV}
\usage{
getPrismaFormat(studyStatus)
}
\arguments{
\item{studyStatus}{a dataframe that keeps track of studies used.}
}
\value{
prismaFormat A dataframe with information on how to make a PRISMA diagram.
}
\description{
Create Prisma
These functions create a PRISMA diagram from a simple dataframe keeping track of sources and filters
It creates a DiagrammeR file run by grV
}
\examples{
prismaTemplate <- getPrisma(studyStatus)
prismaFormat <- getPrismaFormat(studyStatus)
prismaTemplate <- getPrisma(studyStatus, prismaFormat)
DiagrammeR::grViz(prismaTemplate)
}
