% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prismaExampleInput}
\alias{prismaExampleInput}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{Source}{Source label to be sorted}
  \item{Source Text}{Source text used}
  \item{Filter}{Filter label used to be sorted}
  \item{Filter Text}{Filter text use to be shown on the side}
}
}
\usage{
prismaExampleInput
}
\description{
A dataset containing one row per entry in the PRISMA diagram.
}
\keyword{datasets}
