% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{studyStatus}
\alias{studyStatus}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 10 rows and 4 variables:
\describe{
  \item{Pub.ID}{Pubblication unique identifier}
  \item{Source}{Source identifier}
  \item{Filter}{Filter identifier}
}
}
\usage{
studyStatus
}
\description{
A dataset containing one row per entry in the PRISMA diagram.
}
\keyword{datasets}
