% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prismadiagramR-package.R
\docType{package}
\name{prismadiagramR-package}
\alias{prismadiagramR}
\alias{prismadiagramR-package}
\title{prismadiagramR: Creates a prisma diagram using DiagrammeR}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Lionel Duarte \email{lduarte@doctorevidence.com} (\href{https://orcid.org/0000-0003-0296-4949}{ORCID})

}
\keyword{internal}
